+++ CHANGELOG.md
@@ -34,10 +34,12 @@
 #### Additions
 
 #### Changes
+* Massive performance improvements for Cargo networks
 
 #### Fixes
 * Fixed #3218
 * Fixed #3241
+* Fixed #3248
 
 ## Release Candidate 28 (06 Sep 2021)